+++ processor/abci.go
@@ -579,6 +579,8 @@ func (app *App) OnCommit() (resp tmtypes.ResponseCommit) {
 	app.log.Debug("Processor COMMIT starting")
 	defer app.log.Debug("Processor COMMIT completed")
 
+	// call checkpoint _first_ so the snapshot contains the correct checkpoint state.
+	cpt, _ := app.checkpoint.Checkpoint(app.blockCtx, app.currentTimestamp)
 	snapHash, err := app.snapshot.Snapshot(app.blockCtx)
 	if err != nil {
 		app.log.Panic("Failed to create snapshot",
@@ -592,9 +594,12 @@ func (app *App) OnCommit() (resp tmtypes.ResponseCommit) {
 		resp.Data = append(resp.Data, app.stakingAccounts.Hash()...)
 	}
 
-	// Checkpoint can be nil if it wasn't time to create a checkpoint
-	if cpt, _ := app.checkpoint.Checkpoint(app.blockCtx, app.currentTimestamp); cpt != nil {
-		resp.Data = append(resp.Data, cpt.Hash...)
+	if cpt != nil {
+		if len(snapHash) == 0 {
+			// only append to commit hash if we aren't using the snapshot hash
+			// otherwise restoring a checkpoint would restore an incomplete/wrong hash
+			resp.Data = append(resp.Data, cpt.Hash...)
+		}
 		_ = app.handleCheckpoint(cpt)
 	}
 	// Compute the AppHash and update the response